Added patch to link against libboost_signal on OSX, in response to #19.
A better fix would be to update the Boost-finding logic, starting with
the CMake version check, which is looking for exactly 2.6.x, which causes
the older logic, intended for 2.4.x, to be used for everything else,
including versions > 2.6.x.  To minimize risk, I instead just added an
OSX-specific clause inside the older logic.